blog/profiles/serializers.py
from rest_framework import serializers
from .models import profile

class profileSerializer(serializers.ModelSerializer):
    owner = serializers.ReadOnlyField(source='owner.username')  # Corrected field name

    class Meta:
        model = profile
        fields = [
            'id', 'owner', 'created_at', 'updated_at', 'name',
            'content', 'image'
        ]
